Add a separate test class for the DistancePIDController
